Add then to list signal
